+++ oshi-core/src/main/java/oshi/software/os/linux/LinuxOperatingSystem.java
@@ -639,4 +639,5 @@ public class LinuxOperatingSystem extends AbstractOperatingSystem {
             return name.substring(0, 1).toUpperCase() + name.substring(1);
         }
     }
+
 }