+++ oshi-core/src/main/java/oshi/software/os/mac/MacOperatingSystem.java
@@ -318,4 +318,5 @@ public class MacOperatingSystem extends AbstractOperatingSystem {
     public NetworkParams getNetworkParams() {
         return new MacNetworkParams();
     }
+
 }