+++ oshi-core/src/main/java/oshi/software/os/unix/freebsd/FreeBsdOperatingSystem.java
@@ -190,4 +190,5 @@ public class FreeBsdOperatingSystem extends AbstractOperatingSystem {
     public NetworkParams getNetworkParams() {
         return new FreeBsdNetworkParams();
     }
+
 }