+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ================
 * [#433] (https://github.com/oshi/oshi/pull/433): Performance improvements for getProcesses() on Linux - [@bildechinger](https://github.com/bildechinger).
 * [#455] (https://github.com/oshi/oshi/pull/455): Open files/handles support - [@spyhunter99](https://github.com/spyhunter99).
+* [#459] (https://github.com/oshi/oshi/pull/459): New methods for querying for a list of specific pids - [@spyhunter99](https://github.com/spyhunter99).
 * Your contribution here.
 
 3.4.4 (10/15/2017)